rename utils/data.py to utils/figshare.py
