+++ src/searchnets/nets/alexnet.py
@@ -6,7 +6,7 @@ import os
 import tensorflow as tf
 import numpy as np
 
-from ..utils.data import fetch
+from ..utils.figshare import fetch
 from ..utils.figshare_urls import ALEXNET_WEIGHTS_URL
 from .layers import max_pool, lrn, dropout
